+++ src/main/java/com/lms/repositories/PrivilegeRepository.java
@@ -7,6 +7,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 /**
  * Created by umit.kas on 28.11.2017.
  */
-public interface PrivilegeRepository extends JpaRepository<Privilege, Integer>{
+public interface PrivilegeRepository extends JpaRepository<Privilege, Long>{
 
 }
