+++ src/main/java/com/lms/repositories/UserCoursePrivilegeRepository.java
@@ -6,5 +6,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 /**
  * Created by umit.kas on 28.11.2017.
  */
-public interface UserCoursePrivilegeRepository extends JpaRepository<UserCoursePrivilege, Integer>{
+public interface UserCoursePrivilegeRepository extends JpaRepository<UserCoursePrivilege, Long>{
 }
